english fixes, move features to generics
