+++ frontend/src/features/Generics/BTypography.js
@@ -1,0 +1,5 @@
+// Custom Typography
+const BTypography = styled(Typography)({
+  textTransform: 'none',
+});
+export default BTypography;
