+++ frontend/src/pages/BusinessManagement.js
@@ -7,7 +7,7 @@ import {
   Typography,
   CircularProgress,
 } from '@mui/material';
-import FrostedBackground from '../features/FrostedBackground';
+import FrostedBackground from '../features/Generics/FrostedBackground';
 import { getAuth, signOut } from 'firebase/auth';
 import { db } from '../firebase';
 import { doc, getDoc, updateDoc } from 'firebase/firestore';
